requirements.txt
dbfread
catboost
category_encoders
chardet
coverage
folium
geopandas
guzzle_sphinx_theme
jyquickhelper>=0.3
mlinsights
msgpack
pandas_streaming
pyensae
pyltr
pyproj
pyquickhelper>=1.8
sphinx-bootstrap-theme
sphinx-gallery
sphinx-redactor-theme
ujson
wheel
xgboost
